+++ backend/src/main/java/com/flavien/sporttracker/service/WorkoutExerciseService.java
@@ -9,9 +9,7 @@ import com.flavien.sporttracker.dto.response.workoutExercise.WorkoutExerciseList
 import com.flavien.sporttracker.entity.WorkoutExercise;
 import com.flavien.sporttracker.exception.customException.NotFoundException;
 import com.flavien.sporttracker.mapper.WorkoutExerciseMapper;
-import com.flavien.sporttracker.repository.ExerciseRepository;
 import com.flavien.sporttracker.repository.WorkoutExerciseRepository;
-import com.flavien.sporttracker.repository.WorkoutRepository;
 
 import lombok.RequiredArgsConstructor;
 
@@ -21,24 +19,20 @@ public class WorkoutExerciseService {
 
         private final WorkoutExerciseRepository workoutExerciseRepository;
 
-        private final WorkoutRepository workoutRepository;
-
-        private final ExerciseRepository exerciseRepository;
-
         private final WorkoutExerciseMapper workoutExerciseMapper;
+
+        private final WorkoutService workoutService;
+
+        private final ExerciseService exerciseService;
 
         public WorkoutExerciseListResponse addWorkoutExerciseById(
                         UUID workoutId,
                         WorkoutExerciseCreateRequest workoutExerciseCreateRequest,
                         String email) {
 
-                var workout = workoutRepository.findByIdAndUserEmail(workoutId, email)
-                                .orElseThrow(() -> new NotFoundException("Workout not found with id: " + workoutId));
-                var exercise = exerciseRepository
-                                .findByIdAndAvailableForUser(workoutExerciseCreateRequest.exerciseId(), email)
-                                .orElseThrow(() -> new NotFoundException(
-                                                "Exercise not found with id: "
-                                                                + workoutExerciseCreateRequest.exerciseId()));
+                var workout = workoutService.getByIdAndEmail(workoutId, email);
+                var exercise = exerciseService.getByIdAndAvailableForUser(workoutExerciseCreateRequest.exerciseId(),
+                                email);
 
                 long maxOrder = workoutExerciseRepository.findMaxOrderByWorkoutId(workoutId).orElse(0L);
 
@@ -53,13 +47,24 @@ public class WorkoutExerciseService {
         }
 
         public void deleteWorkoutExerciseById(UUID workoutId, UUID workoutExerciseId, String email) {
-                workoutRepository.findByIdAndUserEmail(workoutId, email)
-                                .orElseThrow(() -> new NotFoundException("Workout not found with id: " + workoutId));
+                workoutService.checkWorkoutOwnership(workoutId, email);
 
                 WorkoutExercise workoutExercise = workoutExerciseRepository
                                 .findByIdAndWorkoutId(workoutExerciseId, workoutId)
                                 .orElseThrow(() -> new NotFoundException("WorkoutExercise not found"));
                 workoutExerciseRepository.delete(workoutExercise);
 
+        }
+
+        public WorkoutExercise getByIdAndWorkoutId(UUID id, UUID workoutId) {
+                return workoutExerciseRepository.findByIdAndWorkoutId(id, workoutId)
+                                .orElseThrow(() -> new NotFoundException("WorkoutExercise not found"));
+        }
+
+        public void checkWorkoutExerciseOwnership(UUID workoutExerciseId, UUID workoutId, String email) {
+                workoutService.checkWorkoutOwnership(workoutId, email);
+                if (!workoutExerciseRepository.existsByIdAndWorkoutId(workoutExerciseId, workoutId)) {
+                        throw new NotFoundException("WorkoutExercise not found with id: " + workoutExerciseId);
+                }
         }
 }
